# Conflicts:
#   frontend/src/components/header/AfterloginMenu.tsx

+++ backend/auction/src/main/java/com/dokidoki/auction/domain/entity/AuctionEndEntity.java
@@ -37,10 +37,10 @@ public class AuctionEndEntity {
     private String description;         // 제품설명
 
     @Column(name = "offer_price")
-    private Integer offerPrice;         // 시작가
+    private Long offerPrice;         // 시작가
 
     @Column(name = "final_price")
-    private Integer finalPrice;         // 낙찰금액
+    private Long finalPrice;         // 낙찰금액
 
     public static AuctionEndEntity createAuctionEnd(
             Long id,
@@ -50,8 +50,8 @@ public class AuctionEndEntity {
             LocalDateTime startTime,
             LocalDateTime endTime,
             String title,
-            Integer offerPrice,
-            Integer finalPrice,
+            Long offerPrice,
+            Long finalPrice,
             String description
     ) {
         AuctionEndEntity auctionEndEntity = new AuctionEndEntity();
