+++ backend/notice/src/main/java/com/dokidoki/notice/api/response/LeaderBoardMemberResp.java
@@ -22,9 +22,9 @@ public class LeaderBoardMemberResp {
     @JsonSerialize(using = LocalDateTimeSerializer.class)
     @JsonDeserialize(using = LocalDateTimeDeserializer.class)
     private LocalDateTime bidTime;
-    private int bidPrice;
+    private Long bidPrice;
 
-    public static LeaderBoardMemberResp of(LeaderBoardMemberInfo info, int bidPrice) {
+    public static LeaderBoardMemberResp of(LeaderBoardMemberInfo info, Long bidPrice) {
         // 개인정보 가리기
         String name = info.getName();
         String modifiedName = name.substring(0, 1) + "*" + name.substring(2);
