+++ backend/notice/src/main/java/com/dokidoki/notice/kafka/dto/KafkaAuctionUpdateDTO.java
@@ -10,13 +10,13 @@ import lombok.ToString;
 @ToString
 public class KafkaAuctionUpdateDTO {
 
-    private long auctionId;
-    private int priceSize;              // 경매 단위
+    private Long auctionId;
+    private Long priceSize;              // 경매 단위
 
     public KafkaAuctionUpdateDTO() {}
 
     @Builder
-    public KafkaAuctionUpdateDTO(long auctionId, int priceSize) {
+    public KafkaAuctionUpdateDTO(Long auctionId, Long priceSize) {
         this.auctionId = auctionId;
         this.priceSize = priceSize;
     }
